Fixed stat editing.
The last changes of adding input features GP, wins, and Loss, caused a bunch of issues with stat editing the edit save function.  I think it should be fixed now with this update.
If you didnt know, before you click predict the stats in the table are editable, you can click and tweak the player stats if you think a player will do good or bad. press enter to save the stat and when you click predict, it rewrites the games dataset csv with the new updated values!