training data player trades and injury
with this update we are now accounting for injuries and mid season trades in training dataset!  players are only added to a game if they scored at least 1 point.  This caused some games to be removed as they didn't have 7 players that scored 1 point... not too worried about it, may reduce players per team to 5 or something.  But this appears to have improved accuracy a good bit, especially combined with the earlier streaks update.  Also about to hit daily rate limit on team data api.