game bets button
there is now a separate view to track game bets.  For now the button is on the profile page next to team list.  the bets page allows for filtering games, set bet to true at the bottom of a game and it will appear in your bets once the game is finished.